nsa infer for module
