+++ native_sparse_attention/infer/__init__.py
@@ -1,0 +1,7 @@
+from native_sparse_attention.infer.inference_func import (
+    compress_infer,
+    compressed_attention_infer,
+    topk_sparse_attention_infer,
+    sliding_window_attention_infer,
+)
+from native_sparse_attention.infer.nsa_inference import nsa_infer
